Add a subdirectory Plotter/EfficiencyPlotterScripts/ containing dedicated python modules for plotting efficiencies in a clever and efficient (no pun intended) way.
